move component files to folder
